week5/1-hello-world/hello_world/src/pyproject.toml
[project]
name = "hello-world"
version = "0.1.0"
description = "The customary start of any learning journey"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.14",
    "pydantic>=2.11.7",
    "strands-agents>=0.1.9",
    "strands-agents-tools>=0.1.7",
]
